Add more JSON parser unit tests and fix found bugs Add some custom unit tests as well as running tests from json.org. To run these tests, needed to add a Path utility to list all files under a directory. Also added a build system feature to pass arguments onto the running target, in order to be able to pass --gtest_filter into the unit tests.
Found some bugs (and unimplemented features) with json.org test files:
1. The parser didn't handle escaped characters inside strings. This is mostly
   complete now, with the exception of \u.
2. The parser allowed non-space whitespace in strings (e.g tabs and newlines).
3. The parser only accepted JSON objects. It should also accept JSON arrays.
4. The parser didn't catch an array with no first object (e.g. [ , 1]);
5. The parser didn't allow commas inside object names, it only allowed them
   inside values.
6. The parser could not handle integer exponentials, e.g. 100E5. Functions like
   std::stoi() do not support exponentials. So instead, convert these strings to
   a float first, then cast to integer.
7. The parser allowed octal numbers, i.e. numbers starting with '0'.

+++ fly/parser/json.h
@@ -70,7 +70,7 @@ public:
     using boolean_type = bool;
     using signed_type = int64_t;
     using unsigned_type = uint64_t;
-    using float_type = double;
+    using float_type = long double;
     using null_type = std::nullptr_t;
 
     /**
